requirements.txt
numpy==1.26.4
molfeat==0.10.0
rdkit==2023.9.6
pandas==2.0.0
torch==2.3.1
torch_geometric==2.5.2
safe-mol==0.1.6
selfies==2.1.1
deepchem==2.8.0
sentence-transformers==3.1.1
transformers==4.39.3
scikit-dimension==0.3.4
scikit-learn==1.5.1
tqdm==4.66.1
matplotlib==3.9.0
seaborn==0.13.2
scipy==1.12.0
persim==0.3.5
giotto-tda==0.6.0
networkx==3.4.2
ripser==0.6.8
rogi==0.1
plotly==5.20.0
altair==5.3.0
statsmodels==0.14.2